Add files needed to be able to run tests on swarming.
In order to migrate off of the 32-bit builder and have
proper tryservers set up, we need to make sure we can
run the tests under swarming, rather than running them
locally.

This CL adds a swarming_client dependency to the DEPS file,
modifies the GYP build and run_tests.py to be able to generate
isolates for swarming, adds a wrapper script to launch
`run_tests` under swarming, and adds a file to describe
which tests to run on swarming on which builders.

Bug: 697636, 743139
Change-Id: Iab3136fc9a323caf368b54b0b8c5192f5cf29477