win: Set PSAPI_VERSION to 2
Upstreaming of https://chromium-review.googlesource.com/c/588209

From MSDN:
"If PSAPI_VERSION is 2 or greater, this function is defined as K32QueryWorkingSet
in Psapi.h and exported in Kernel32.lib and Kernel32.dll. If PSAPI_VERSION is 1,
this function is defined as QueryWorkingSet in Psapi.h and exported in Psapi.lib
and Psapi.dll as a wrapper that calls K32QueryWorkingSet. Programs that must run
on earlier versions of Windows as well as Windows 7 and later versions should
always call this function as QueryWorkingSet. To ensure correct resolution of
symbols, add Psapi.lib to the TARGETLIBS macro and compile the program with
-DPSAPI_VERSION=1. To use run-time dynamic linking, load Psapi.dll."

Therefore, we should switch to PSAPI_VERSION==2 then we shouldn't need psapi.lib
and psapi.dll anymore.

Change-Id: Id5d601d9880dcf901b9877afa6712d43bf1d4119