Create client data structures for typed Annotations.
This introduces the Annotation object, used to declare typed annotations,
and the AnnotationList object, used to reference these. The AnnotationList
is referenced by the CrashpadInfo structure. Currently nothing reads these.

The AnnotationList implements a lock-free linked list, into which Annotation
objects are added exactly once, when they are first set. Clearing an Annotation
merely marks it internally as such, rather than removing it from the list.

Bug: crashpad:192
Change-Id: I72414b1f83d624c4ae323e09ecea8cfb69a68c5e