Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Thanks for the comments, Mark. I'm going to back off on this a little, moving the {0} changes to a later CL.
> 
> For the others (nullptr_t override, offsetof() include): Is there a recommended (or required-minimum) toolchain to build with on desktop Linux?

I’m mainly working with trunkish clang. I’ve also been trying to make sure that recent (6.3-vintage) GCC works. Linux and Android clients are still in development, so we haven’t nailed anything down very firmly yet. The only hard rule for the time being is that we need to be able to build with the compiler in the current NDK, which is LLVM 5.0 clang. Further down the road, the rule will be “whatever compiler Chrome is using” which for most platforms is trunkish clang.

As you progress, if you have particular toolchain requirements, I’m happy to enforce those too in the form of try/buildbots. In fact, if you have any targets at all building now, it’s probably not too early to think about setting this up.

> It looks like ninja uses plain "c++", so I'm at the whim of my system.
> 
> I originally ran into the nullptr_t override issue when building with the g++ that comes with Ubuntu 14.04.5.

So probably 4.8-vintage, then. That’s pretty old now…

> I recently upgraded my machine to a new-ish Debian testing build, and the installed c++ no longer needs that override, although it still wants <stddef.h> to define offsetof(). My default c++ is now x86_64-linux-gnu-g++-6, version (Debian 6.3.0-18) 6.3.0 20170516.
> 
> And maybe the offsetof() problem would go away if I were using the same toolchain that the crashpad team is using.

Maybe, but it wouldn’t be correct. We do want <stddef.h> wherever offsetof() is used. Even if that’s all that comes of this patch, I think we should take it.

Patch-set: 2
Reviewer: Mark Mentovai <1174099@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>