Update patch set 3
Patch Set 3: Code-Review+1

If this works most of the time, but not all of the time, with gtest’s SEH handler, but works all of the time when we switch to this, then we may need to look at what gtest is doing in its handler.

Provisional LGTM, so we can see whether anything changes.

Patch-set: 3
Reviewer: Mark Mentovai <1174099@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1