win handler: Provide a wmain() entry point
Crashpad’s own build always uses wWinMain(), the default entry point for
/subsystem:windows, producing crashpad_handler.exe. A /subsystem:console
version, crashpad_handler.com, is also produced, and it also uses
wWinMain() and not the traditional wmain() used in /subsystem:console,
because it is produced by using editbin on a copy of
crashpad_handler.exe to change just the subsystem, leaving the entry
point intact.

For the in-Chromium build’s tests, it is conveient to produce the
/subsystem:console version, crashpad_handler.com, directly as build
output, as opposed to using editbin to transform a /subsystem:windows
version. This /subsystem:console version uses the normal wmain() entry
point.

By providing both wWinMain() and wmain(), both build types can be
accommodated.

Bug: chromium:779790
Change-Id: Ieb784db0cc245c6e4c12fb1dd83b8b95e159bdec