Update patch set 7
Patch Set 7:

> thanks for tackling this, looks pretty reasonable to my eye. Testing is going to be a PITA, as it always is for forwards/backwards compatibility.

The plan is to have one module with a “small” (old) CrashpadInfo struct and one with a “large” one.

But first, I’m taking a look at the failures from patch set 5.

https://build.chromium.org/p/client.crashpad/builders/crashpad_try_win_x86_wow64_dbg/builds/573/steps/run%20tests/logs/stdio
https://build.chromium.org/p/client.crashpad/builders/crashpad_try_win_x64_dbg/builds/563/steps/run%20tests/logs/stdio

When I was looking for strict size equality between the section and struct, I got this on the bots (but not locally):

[pid:tid:yyyymmdd,hhmmss.mmm:WARNING pe_image_reader.cc:116] crashpad info struct size 44 mismatch with section size 308 in E:\b\build\slave\crashpad_try_win_x86_wow64_dbg\build\crashpad\out\Debug\crashpad_snapshot_test.exe

That was 32-bit. 64-bit, the numbers were 64 and 332. 308÷44 is exactly 7, although I can’t make up 332 using just 64s (or 64s and 44s).

In patch set 7, I’m back to just checking that the struct size fits in the section, but now I’m worried that there are extra structs going into the section, and I want to figure that out before moving forward.

I couldn’t reproduce locally, but the bots are building with MSVS 2015 instead of 2017, so I’m going to try that next.

Also, I’m clicking “start review” because that seems to be the only way to send a message and reply to you, but this is still “work-in-progress” as far as I’m concerned.

Patch-set: 7
Reviewer: Mark Mentovai <1174099@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0
Work-in-progress: false