Update patch set 21
Patch Set 21:

OK, this seems like a plausible not-too-ugly checkpoint to land and make sure it doesn't break Chromium, PTAL.

I'm not 100% sure about the empty-files-in-special-places (mostly because it makes you wonder what's happening when you look at a real Crashpad build file) but it avoids explicit "I'm in Chrome" flags that I had in early patchsets, which is nice.

With one local change on top of this to third_party/zlib/BUILD.gn to not refer to itself, gn generation completes and generates something that ninja complains about a bit, but then tries to build most targets. And fails.

Patch-set: 21
Reviewer: Scott Graham <1144260@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Work-in-progress: false