Update patch set 7
Patch Set 7:

Based on the bug reports, it sounds like there may be restrictions on what goes where in the filesystem on Fuchsia. Is that true? Seems like a lot of containerizing and sandboxing, which is good.

What I'm getting at is that this interface may not even be sensible on Fuchsia at all if the executable's always supposed to show up at the same path in some container, and test data would have its own proper home. What do you think? I've got "try to run it" on my list of things to do, so I'm not entirely sure of any of this yet, but it's something that occurred to me.

(Haven't actually looked at the latest patch set yet, I'm just talking out loud.)

Patch-set: 7
Reviewer: Mark Mentovai <1174099@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>