Update patch set 7
Patch Set 7:

This is getting way complicated, mostly unrelated to this CL -- I need to have all the setup gewgaws that allow packaging the result of gn runtime deps to create a package that can be transported to a target device (and then I can to the namespace mapping thing described at https://fuchsia.googlesource.com/docs/+/master/namespaces.md#typical-directory-structure). So this CL will be on hold for a while until that's sorted out.

It seems like we'd need that for Android too since it's not self-hosting. Has anyone looked at that already?

Patch-set: 7
Reviewer: Scott Graham <1144260@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>