fuchsia: Get 'all' to build
Adds a standalone zlib build file when building standalone. This
avoids including the code for minizip, etc. (instead, only the main
libzlib.a static_library). The other libraries and executables refer
to zlib includes and so don't build; having our own build file means
that all defined targets in the GN build are buildable.

generate_dump is passing an invalid handle to ProcessSnapshotFuchsia
as there's no utility to convert a pid to a handle. But that's no
great loss, because ProcessSnapshotFuchsia doesn't do anything itself.

Bug: crashpad:79, crashpad:196
Change-Id: I11c918a30b60cc071465c919315b45caab1de870