win: Fix ProcessInfo.OtherProcess test flake
Removes the /BASE:N and /FIXED arguments to the child, which
weren't actually testing correctly (see bug), and were causing
problems at least on Win7 when something collided with that address.

Additionally, switches to storing modules in load order, rather than
a combination of memory order and initialization order, since that
was a bit confusing and there was no great rationale for it.

Bug: chromium:792619
Change-Id: I2e95a81b02fe4d527868f6a5f980d315604255a6