Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > > Patch Set 1: Commit-Queue+2
> > > 
> > > It was supposed to be a holiday gift for you! I saved the gift receipt if you want to exchange it for something else.
> > 
> > I'll exchange it for the purpley-red-coloured version instead!
> 
> More because I was interested in the “how” on purpley-red, I threw https://chromium-review.googlesource.com/c/820532/ together.
> 
> It doesn’t do anything useful (yet?), but it shows how it could be done.
> 
> Happy Boxing Day!

Heh :)

I was also investigating, via file redirection to a local file on the device, and then netcp and printing that, which seems to work ok too. https://chromium-review.googlesource.com/#/c/crashpad/crashpad/+/820610

But then I got sidetracked into hunting down where in Fuchsia the loss of the \x1b actually happens because it seems like it shouldn't if I tell it what TERM I want to pretend I'm using... it seems like most things are doing a straight memcpy, but something! is obviously doing it.

Patch-set: 2