Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > > Patch Set 2:
> > > 
> > > > Patch Set 1: Commit-Queue+2
> > > > 
> > > > It was supposed to be a holiday gift for you! I saved the gift receipt if you want to exchange it for something else.
> > > 
> > > I'll exchange it for the purpley-red-coloured version instead!
> > 
> > More because I was interested in the “how” on purpley-red, I threw https://chromium-review.googlesource.com/c/820532/ together.
> > 
> > It doesn’t do anything useful (yet?), but it shows how it could be done.
> > 
> > Happy Boxing Day!
> 
> Heh :)
> 
> I was also investigating, via file redirection to a local file on the device, and then netcp and printing that, which seems to work ok too. https://chromium-review.googlesource.com/#/c/crashpad/crashpad/+/820610
> 
> But then I got sidetracked into hunting down where in Fuchsia the loss of the \x1b actually happens because it seems like it shouldn't if I tell it what TERM I want to pretend I'm using... it seems like most things are doing a straight memcpy, but something! is obviously doing it.

I eventually tracked it down to:
https://fuchsia.googlesource.com/zircon/+/HEAD/system/ulib/fdio/logger.c#49
which is from the initial commit, and so has no rationale. I guess it's probably intended to avoid accidental ansi art in a hopefully linear output log. So, not too useful of a finding as I likely can't remove it, but ... I guess I got a tour of Fuchsia IO at least.

It's sort of at the wrong layer (i.e. fdio, not the log itself) so printfs/puts will drop escapes, but zx_log_write() won't. So I think the wrapper could simply be a cat-alike that sends stdin to zx_log_write(), and then the host side could just continue using loglistener as it does now.

Patch-set: 2