Update patch set 8
Patch Set 8:

If this gets too hairy, it looks (based on https://code.google.com/p/android/issues/detail?id=61799) that we can avoid the crash by not specifying RTLD_DEFAULT, but instead using a dlopened handle to the library known to host the symbol we’re interested in.

Usually when I do this, I like to dladdr a nearby symbol and use the returned dli_fname as the target of my dlopen. “git grep” for dladdr and you’ll find a couple of examples. Normally this is how I do things on macOS, to match the usual “two-level namespace” semantics that you get there by default. On ELF platforms like Linux, I do usually just use RTLD_DEFAULT, because that’s a better match for what normally happens on those platforms. So in this case, if you want to head down this suggested path, you might want to version-check it and only do the dladdr-dlopen thing if you’re on a version known to be broken, and keep using RTLD_DEFAULT everywhere else.

Patch-set: 8
Reviewer: Mark Mentovai <1174099@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>