Update patch set 8
Patch Set 8:

> Patch Set 8:
> 
> > Yes! RTLD_NEXT.
> > 
> > It should be safe even on K because we always expect dlsym to be defined.
> 
> Oh, but you can't call dlsym in the first place...
> 
> I recall that there's some linker magic that can be done here. I'll look a little later. But this may also be another reason to go with the simplified suggestion of using a real dl handle instead of RTLD_DEFAULT on K, rather than overriding anything.

An option is “ld --wrap dlsym”, which will replace references to dlsym with ones to __wrap_dlsym, while making the original dlsym appear as __real_dlsym. Then, we can write __wrap_dlsym(), which calls __real_dlsym(). __wrap_dlsym() will need to be extern "C", and we’ll need an extern "C" declaration for __real_dlsym(). This will work, but GN makes it difficult to push the ldflags to where we need them (https://crbug.com/796183), and if we fail to do so, dlsym() callers will silently get libdl.so’s dlsym() without any wrapping, exactly what we’re trying to avoid.

So we’ve got that option, and we’ve got the dl-handle-instead-of-RTLD_DEFAULT-on-K option. Finally, we could just give up on wrapping dlsym by that name as an interface that’s exposed from compat, and concede by doing it as a wrapper by another name (as you’ve already done). We wouldn’t want it to be exposed as an interface from compat, so that’d mean not declaring it in a “public” header (we don’t have real header visibility, but for compat, we should consider anything that’s shadowing or filling in for a system header to be public, so <dlfcn.h> is off limits for an internal function), #including it only from .cc files, maybe using an internal namespace, etc. Since it’s our own interface, it’ll free us from having to do dlerror() compatibility, letting you focus on just the meat of the siglongjmp().

Patch-set: 8
Reviewer: Mark Mentovai <1174099@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>