Update patch set 2
Patch Set 2: Code-Review+1

Oh, this is unfortunate.

And it’s not available under any other name, or something that we can dlsym() for?

I suppose that for a test, it’s not really worth providing our own fallback implementation since it could just devolve to another copy of the code that we’re trying to test. Not too useful…

Patch-set: 2
Reviewer: Mark Mentovai <1174099@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1